requirements.txt
elasticsearch
docker
sigmatools
click
pyyaml